+++ docs/source/changelog.rst
@@ -14,6 +14,7 @@ History
 Version 0.2.28 (2022-09-13)
 ---------------------------
 
+* Fixed bug in Node repr if operator is None.
 * Added new ``population_growth_function`` to Island, easier to switch between stead-state and generational.
 * Added new population growth functions to ``utils``.